+++ ktorfit-lib-common/src/commonMain/kotlin/de/jensklingenberg/ktorfit/TypeInfoExt.kt
@@ -9,6 +9,7 @@ import kotlin.reflect.KClass
  * Example: Response<String> will return String as TypeInfo with upperBoundType(0)
  */
 public fun TypeInfo.upperBoundType(index: Int = 0): TypeInfo? {
+
     val parentType = this.kotlinType ?: return null
     val modelKTypeProjection = if (parentType.arguments.isNotEmpty()) parentType.arguments[index] else return null
     val modelKType = modelKTypeProjection.type ?: return null