+++ ktorfit-lib-common/src/commonMain/kotlin/de/jensklingenberg/ktorfit/converter/builtin/CallConverterFactory.kt
@@ -78,3 +78,4 @@ public class CallConverterFactory : Converter.Factory {
         return null
     }
 }
+