+++ ktorfit-lib-common/src/commonMain/kotlin/de/jensklingenberg/ktorfit/internal/TypeData.kt
@@ -1,6 +1,7 @@
 package de.jensklingenberg.ktorfit.internal
 
 import io.ktor.util.reflect.*
+import kotlin.reflect.KClass
 
 /**
  * This class contains information about the requested response type
@@ -17,7 +18,22 @@ public data class TypeData(
     public val isNullable: Boolean = qualifiedName.endsWith("?"),
     val typeInfo: TypeInfo,
 ){
-    internal companion object{
+    public companion object{
+        public fun createTypeData(qualifiedTypename: String, typeInfo: TypeInfo): TypeData {
 
+            val typeArgument = qualifiedTypename.substringAfter("<").substringBeforeLast(">")
+            val spilt = typeArgument.split(",")
+            val args = mutableListOf<TypeData>()
+            typeInfo.kotlinType?.arguments?.forEachIndexed { index, kTypeProjection ->
+                val cleaned = spilt[index].trim()
+
+                val modelKType = kTypeProjection.type
+                val modelClass = (modelKType?.classifier as? KClass<*>?)!!
+
+                args.add(createTypeData(cleaned, TypeInfo(modelClass, modelKType.platformType, modelKType)))
+            }
+
+            return TypeData(qualifiedTypename.substringBefore("<"), args, typeInfo = typeInfo)
+        }
     }
 }